requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
email-validator==2.1.0
Pillow==10.0.1
WTForms==3.1.0
python-dotenv==1.0.0 